api_db/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
fdb==2.0.4
python-dotenv==1.0.0
